c# multiples-of-3-or-5
A very language agnostic kata. The real solution is a tiny bit of math.

The naive solution is the simple loop and checking every value 1..n.
The smarter solution is to use stepsizes 3 and 5 in seperate loops.
And the best solution is simply to calculate with formulas and not
iterate at all!